--- a/node/src/service.rs
+++ b/node/src/service.rs
@@ -570,7 +570,7 @@ pub fn new_full_base(
 			force_authoring,
 			backoff_authoring_blocks,
 			babe_link,
-			block_proposal_slot_portion: SlotProportion::new(0.25) // changeed from 0.5, slower production rate,
+			block_proposal_slot_portion: SlotProportion::new(0.25), // changeed from 0.5, slower production rate,
 			max_block_proposal_slot_portion: None,
 			telemetry: telemetry.as_ref().map(|x| x.handle()),
 		};