Modified consensus parameters to improve finalization rate and reduce gap between block production and finalization:
- Increased GRANDPA gossip duration (333ms → 1000ms) for better vote propagation
- Reduced justification period (512 → 256) for faster finality confirmation  
- Increased block time (6s → 12s) to allow more time for finalization
- Adjusted BABE probability (1/4 → 1/3) and reduced epoch duration for better block production pacing
- Optimized session length (6 → 4) for more frequent validator updates

These changes aim to prevent block production from overtaking finalization while maintaining network stability.